Added train generator
Added a train generator. Also, a way to convert expressions to tokens in order to include poly parameters in the tokenization procedure, and a way to convert tokens to expressions. The image generated is the graph and the file name of each graph can be used as token inputs for training.